+++ R/learner_xgboost_surv_xgboost_cox.R
@@ -40,6 +40,16 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
+      p_nrounds = p_int(1L,
+        tags = c("train", "hotstart", "internal_tuning"),
+        aggr = crate(function(x) as.integer(ceiling(mean(unlist(x)))), .parent = topenv()),
+        in_tune_fn = crate(function(domain, param_vals) {
+          if (is.null(param_vals$early_stopping_rounds)) {
+            stop("Parameter 'early_stopping_rounds' must be set to use internal tuning.")
+          }
+          assert_integerish(domain$upper, len = 1L, any.missing = FALSE)}, .parent = topenv()),
+        disable_in_tune = list(early_stopping_rounds = NULL)
+      )
       ps = ps(
         alpha                       = p_dbl(0, default = 0, tags = "train"),
         base_score                  = p_dbl(default = 0.5, tags = "train"),
@@ -50,7 +60,6 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
         colsample_bytree            = p_dbl(0, 1, default = 1, tags = "train"),
         disable_default_eval_metric = p_lgl(default = FALSE, tags = "train"),
         early_stopping_rounds       = p_int(1L, default = NULL, special_vals = list(NULL), tags = "train"),
-        early_stopping_set          = p_fct(c("none", "train", "test"), default = "none", tags = "train"),
         eta                         = p_dbl(0, 1, default = 0.3, tags = "train"),
         feature_selector            = p_fct(c("cyclic", "shuffle", "random", "greedy", "thrifty"), default = "cyclic", tags = "train"), #nolint
         feval                       = p_uty(default = NULL, tags = "train"),
@@ -69,7 +78,7 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
         missing                     = p_dbl(default = NA, tags = c("train", "predict"), special_vals = list(NA, NA_real_, NULL)), #nolint
         monotone_constraints        = p_int(-1L, 1L, default = 0L, tags = "train"),
         normalize_type              = p_fct(c("tree", "forest"), default = "tree", tags = "train"),
-        nrounds                     = p_int(1L, tags = "train"),
+        nrounds                     = p_nrounds,
         nthread                     = p_int(1L, default = 1L, tags = c("train", "threads")),
         num_parallel_tree           = p_int(1L, default = 1L, tags = "train"),
         one_drop                    = p_lgl(default = FALSE, tags = "train"),
@@ -111,7 +120,7 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
       ps$add_dep("top_k", "feature_selector", CondAnyOf$new(c("greedy", "thrifty")))
 
       # custom defaults
-      ps$values = list(nrounds = 1L, nthread = 1L, verbose = 0L, early_stopping_set = "none")
+      ps$values = list(nrounds = 1L, nthread = 1L, verbose = 0L)
 
       super$initialize(
         id = "surv.xgboost.cox",
@@ -134,29 +143,65 @@ LearnerSurvXgboostCox = R6Class("LearnerSurvXgboostCox",
     }
   ),
 
+  active = list(
+    #' @field internal_valid_scores
+    #' The last observation of the validation scores for all metrics.
+    #' Extracted from `model$evaluation_log`
+    internal_valid_scores = function() {
+      self$state$internal_valid_scores
+    },
+    #' @field internal_tuned_values
+    #' Returns the early stopped iterations if `early_stopping_rounds` was set during training.
+    internal_tuned_values = function() {
+      self$state$internal_tuned_values
+    },
+    #' @field validate
+    #' How to construct the internal validation data. This parameter can be either `NULL`,
+    #' a ratio, `"test"`, or `"predefined"`.
+    validate = function(rhs) {
+      if (!missing(rhs)) {
+        private$.validate = assert_validate(rhs)
+      }
+      private$.validate
+    }
+  ),
+
   private = list(
+    .validate = NULL,
+    .extract_internal_tuned_values = function() {
+      if (is.null(self$state$param_vals$early_stopping_rounds)) {
+        return(NULL)
+      }
+      list(nrounds = self$model$model$niter)
+    },
+
+    .extract_internal_valid_scores = function() {
+      if (is.null(self$model$model$evaluation_log)) {
+        return(named_list())
+      }
+      as.list(self$model$model$evaluation_log[
+        get(".N"),
+        set_names(get(".SD"), gsub("^test_", "", colnames(get(".SD")))),
+        .SDcols = patterns("^test_")
+      ])
+    },
     .train = function(task) {
       pv = self$param_set$get_values(tags = "train")
       # manually add 'objective' and 'eval_metric'
       pv = c(pv, objective = "survival:cox", eval_metric = "cox-nloglik")
 
-      data = get_xgb_mat(task, objective = pv$objective)
+      data = get_xgb_mat(task, pv$objective)
 
-      if ("weights" %in% task$properties) {
-        xgboost::setinfo(data, "weight", task$weights$weight)
+      internal_valid_task = task$internal_valid_task
+      if (!is.null(pv$early_stopping_rounds) && is.null(internal_valid_task)) {
+        stopf("Learner (%s): Configure field 'validate' to enable early stopping.", self$id)
       }
-
-      # XGBoost uses the last element in the watchlist as
-      # the early stopping set
-      if (pv$early_stopping_set != "none") {
-        pv$watchlist = c(pv$watchlist, list(train = data))
-      }
-
-      if (pv$early_stopping_set == "test" && !is.null(task$row_roles$test)) {
-        test_data = get_xgb_mat(task, pv$objective, task$row_roles$test)
+      if (!is.null(internal_valid_task)) {
+        test_data = get_xgb_mat(internal_valid_task, pv$objective)
+        # XGBoost uses the last element in the watchlist as
+        # the early stopping set
         pv$watchlist = c(pv$watchlist, list(test = test_data))
       }
-      pv$early_stopping_set = NULL
 
       list(
         model = invoke(xgboost::xgb.train, data = data, .args = pv),